Allow interrupted downloads to be resumed
Now when the launcher gets an error during downloading, it will keep all the blobs that were successfully downloaded. This means restarting the download won't have to redownload that data.

There were basically two problems that needed solving to implement this.

The first problem is that the launcher wants all content blobs to be referenced, or else it'll cull them on update. This is easy enough to deal with: we just track interrupted downloads with some new DB tables and make the culling logic aware of those.

The second problem is keeping state consistent. We can't practically keep the entire database consistent during the download process, so we relied solely on the SQLite transaction to undo everything if something went wrong. Given that we now want to persist the content blobs, that won't do.

Now, if an error gets caught during downloads, we still commit the database but manually delete the broken data. Given that this is obviously a dangerous situation to be in, I rewrote much of the logic to avoid writing invalid data. For example, previously we wrote the content manifest directly to the DB as it was getting read in, which also meant immediately writing empty placeholder blobs. Now the blobs just get added as they're downloaded. All things considered this cleans up the code too so it's probably a good thing.

At first, I wanted to make all downloads first be stored as an interrupted download, and then swap the data out to a proper content version later. I ended up not doing this however as it'd make the joinery between the delta and zip download system more complicated. It would allow us to do intermittent commits, I guess, but it's something to consider for later.

Related to #187, but does not close it (ideally we'd automatically retry for transient network errors).